% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_functions.R
\name{email_send}
\alias{email_send}
\title{Send an Outlook email. Wrapper around email_draft()}
\usage{
email_send(
  to,
  subject,
  body,
  from = NA,
  attach = c(),
  cc = c(),
  bcc = c(),
  visible = F,
  check_ooo = F,
  send = T,
  signature = F
)
}
\arguments{
\item{to}{Main recipients}

\item{subject}{Email subject}

\item{body}{Email body. Searches for html line break to decide if body is html}

\item{from}{Alternate email to send on behalf of}

\item{attach}{Character vector of files to attach}

\item{cc}{Email cc}

\item{bcc}{Email bcc}

\item{visible}{Open the email in the viewer}

\item{check_ooo}{Check for and remove recipients who are out of office the entire day.
If there are no more main recipients, the email is not sent. Requires open Outlook application.}

\item{send}{Send the email}

\item{signature}{Attach signature to end of email. Requires open Outlook application.}
}
\value{
T/F
}
\description{
Send an Outlook email. Wrapper around email_draft()
}
