% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_functions.R
\name{email_setup}
\alias{email_setup}
\title{Checks and walkthrough to enable RDCOMClient emails.}
\usage{
email_setup(force = F, repo = "bschamberger")
}
\arguments{
\item{force}{Ignore previous RDCOMClient installation}

\item{repo}{Github repository to use. omegahat or bschamberger.}
}
\value{
T/F
}
\description{
Checks and walkthrough to enable RDCOMClient emails.
}
